+++ colrcv/models/hsv.c
@@ -21,10 +21,10 @@
 extern "C"{
 #endif
 
-const double COLRCV_HSV_MIN_VALUE = 0;
-const double COLRCV_HSV_H_MAX_VALUE = 360;
-const double COLRCV_HSV_S_MAX_VALUE = 100;
-const double COLRCV_HSV_V_MAX_VALUE = 100;
+const long double COLRCV_HSV_MIN_VALUE = 0;
+const long double COLRCV_HSV_H_MAX_VALUE = 360;
+const long double COLRCV_HSV_S_MAX_VALUE = 100;
+const long double COLRCV_HSV_V_MAX_VALUE = 100;
 
 bool colrcv_hsv_h_is_valid(colrcv_hsv_t hsv) {
     return colrcv_range_valid(
@@ -56,8 +56,8 @@ bool colrcv_hsv_is_valid(colrcv_hsv_t hsv) {
 // Algorithm: http://www.easyrgb.com/index.php?X=MATH&H=21#text21
 colrcv_result_t colrcv_hsv_to_rgb(colrcv_hsv_t hsv, colrcv_rgb_t* rgb) {
     // down-scale s and v input values first
-    const double s = hsv.s / 100;
-    const double v = hsv.v / 100;
+    const long double s = hsv.s / 100;
+    const long double v = hsv.v / 100;
     // if saturation is 0, then we can return early
     if(s == 0) {
         rgb->r = v * 255;
@@ -66,7 +66,7 @@ colrcv_result_t colrcv_hsv_to_rgb(colrcv_hsv_t hsv, colrcv_rgb_t* rgb) {
         return;
     } else {
         // scale down H to be in range 0 -> 6
-        double scaled_h = hsv.h / 60;
+        long double scaled_h = hsv.h / 60;
         // h value must be snapped to range 0.0 -> 6.0
         if(scaled_h >= 6.0) {
             scaled_h -= 6.0;
@@ -74,9 +74,9 @@ colrcv_result_t colrcv_hsv_to_rgb(colrcv_hsv_t hsv, colrcv_rgb_t* rgb) {
         // floor convert to int
         const uint16_t integer_h = (uint16_t)scaled_h;
         // this further lot of temporaries are used in the channel multiplex
-        const double temp_a = v * (1.0 - s);
-        const double temp_b = v * (1.0 - s * (scaled_h - integer_h));
-        const double temp_c = v * (1.0 - s * (1.0 - (scaled_h - integer_h)));
+        const long double temp_a = v * (1.0 - s);
+        const long double temp_b = v * (1.0 - s * (scaled_h - integer_h));
+        const long double temp_c = v * (1.0 - s * (1.0 - (scaled_h - integer_h)));
         // choose one of many different multiplexes based on value of integer_h
         switch(integer_h % 6) {
             case 0:
